Add bounding box and bounding polygon editor for ISO19139/119 records.  * See http://trac.osgeo.org/geonetwork/wiki/GeoEditorViewer for more details Dev's note:  * In order to post fragment of XML to a metadata record in the DataManager, a new prefixed had been added. All elements with a _X will be added as XML to the corresponding element (instead of KVP). In the geoeditor, the client side create a gml:polygon in a hidden form field which is post to the DataManager and then inserted in the element.  * All XSL templates related to extent for ISO19139/119 moved to metadata-iso19139-geo.xsl Forced spatial index prj file to be WGS84. Added some comments in how to spatially index bounding polygons.
git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@5173 ff403467-3f20-0410-99e5-ff581e626a08

+++ src/org/fao/geonet/kernel/search/SearchManager.java
@@ -503,8 +503,8 @@ public class SearchManager
 		// if rebuild forced or bad index then rebuild index
 		if (rebuild || badIndex) {
 			Log.error(Geonet.SEARCH_ENGINE, "Rebuilding lucene index");
-
-			_spatial.writer().reset();
+			if (_spatial != null)
+				_spatial.writer().reset();
 			IndexWriter writer = new IndexWriter(_luceneDir, _analyzer, true);
 			writer.close();
 		}
@@ -677,7 +677,8 @@ public class SearchManager
             // the file
             // and therefore the test will not be worthwhile
             if (!SpatialIndexWriter.createDataStoreFile(appPath).exists()) {
-                rebuildIndex = true;
+            	Log.error(Geonet.SEARCH_ENGINE, "Rebuild index because spatial index does not exist.");
+            	rebuildIndex = true;
             }
             rebuildIndex = createWriter(appPath);
             if (rebuildIndex) {
@@ -699,7 +700,8 @@ public class SearchManager
                         _transaction, _lock);
                 rebuildIndex = _writer.getFeatureSource().getSchema() == null;
             } catch (Exception e) {
-                _writer.delete();
+                if (_writer != null)
+                	_writer.delete();
                 rebuildIndex = true;
             }
             return rebuildIndex;
